+++ puffin-imgui/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## Unreleased
+
+
+## 0.14.0 - 2021-11-12
 * Add slider for controlling number of frames recorded.
 * Show total frames recorded and their total size.
 * Lower the default number of recorded frames to 600.